Problem 2

+++ problems/problem_2.cpp
@@ -1,5 +1,5 @@
 float problemSolution2(float i, float j, float k) {
-    float result;
+    int result;
     // write your code here;
     std::cin >> i>>j>>k;
     result = (i > j) ? ((i > k) ? i : k) : ((j > k) ? j : k);
